+++ common/src/main/java/com/lowdragmc/photon/client/gameobject/particle/BeamParticle.java
@@ -60,6 +60,7 @@ public class BeamParticle implements IParticle {
     }
 
     public void setup() {
+        this.setDelay(config.getStartDelay());
         update();
         updateOrigin();
     }