+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## ChangeLog
 
-* Fixed beam particle and trail particle doesn't have a correct T value.
+* Added beam and trail emitters lifetime.